Reuse common error conversion logic
