+++ chasm/lib/nexusoperation/operation_tasks.go
@@ -24,6 +24,112 @@ import (
 	"go.uber.org/fx"
 )
 
+// startResult is a marker interface for the outcome of a Nexus start operation call.
+type startResult interface {
+	mustImplementStartResult()
+}
+
+// startResultOK indicates the operation completed synchronously or started asynchronously.
+type startResultOK struct {
+	response *nexusrpc.ClientStartOperationResponse[*commonpb.Payload]
+	links    []*commonpb.Link
+}
+
+func (startResultOK) mustImplementStartResult() {}
+
+// startResultFail indicates a non-retryable failure.
+type startResultFail struct {
+	failure *failurepb.Failure
+}
+
+func (startResultFail) mustImplementStartResult() {}
+
+// startResultRetry indicates a retryable failure.
+type startResultRetry struct {
+	failure *failurepb.Failure
+}
+
+func (startResultRetry) mustImplementStartResult() {}
+
+// startResultCancel indicates the operation completed as canceled.
+type startResultCancel struct {
+	failure *failurepb.Failure
+}
+
+func (startResultCancel) mustImplementStartResult() {}
+
+// startResultTimeout indicates the operation timed out while attempting to invoke.
+type startResultTimeout struct {
+	failure *failurepb.Failure
+}
+
+func (startResultTimeout) mustImplementStartResult() {}
+
+func newStartResult(
+	response *nexusrpc.ClientStartOperationResponse[*commonpb.Payload],
+	callErr error,
+) (startResult, error) {
+	if callErr == nil {
+		return startResultOK{response: response}, nil
+	}
+
+	if opErr, ok := errors.AsType[*nexus.OperationError](callErr); ok {
+		failure, err := operationErrorToFailure(opErr)
+		if err != nil {
+			return nil, err
+		}
+		if opErr.State == nexus.OperationStateCanceled {
+			return startResultCancel{failure: failure}, nil
+		}
+		return startResultFail{failure: failure}, nil
+	}
+
+	if opTimeoutBelowMinErr, ok := errors.AsType[*operationTimeoutBelowMinError](callErr); ok {
+		failure := &failurepb.Failure{
+			Message: "operation timed out",
+			FailureInfo: &failurepb.Failure_TimeoutFailureInfo{
+				TimeoutFailureInfo: &failurepb.TimeoutFailureInfo{
+					TimeoutType: opTimeoutBelowMinErr.timeoutType,
+				},
+			},
+		}
+		return startResultTimeout{failure: failure}, nil
+	}
+
+	failure, retryable, err := callErrorToFailure(callErr)
+	if err != nil {
+		return nil, err
+	}
+	if retryable {
+		return startResultRetry{failure: failure}, nil
+	}
+	return startResultFail{failure: failure}, nil
+}
+
+// operationErrorToFailure converts a Nexus OperationError to the appropriate failure.
+func operationErrorToFailure(opErr *nexus.OperationError) (*failurepb.Failure, error) {
+	var nf nexus.Failure
+	if opErr.OriginalFailure != nil {
+		nf = *opErr.OriginalFailure
+	} else {
+		var err error
+		nf, err = nexusrpc.DefaultFailureConverter().ErrorToFailure(opErr)
+		if err != nil {
+			return nil, err
+		}
+	}
+	// Special marker for Temporal->Temporal calls to indicate that the original failure should be unwrapped.
+	// Temporal uses a wrapper operation error with no additional information to transmit the OperationError over the network.
+	// The meaningful information is in the operation error's cause.
+	unwrapError := nf.Metadata["unwrap-error"] == "true"
+
+	if unwrapError && nf.Cause != nil {
+		return commonnexus.NexusFailureToTemporalFailure(*nf.Cause)
+	}
+	// Transform the OperationError to either ApplicationFailure or CanceledFailure based on the operation error state.
+	return commonnexus.NexusFailureToTemporalFailure(nf)
+}
+
 // operationInvocationTaskHandlerOptions is the fx parameter object for the invocation task executor.
 type operationInvocationTaskHandlerOptions struct {
 	fx.In
